Ready for real refactoring
