+++ spec/fn.spec.js
@@ -1,0 +1,5 @@
+import fn from '../src/js/fn';
+
+describe('fn.js', () => {
+    fn;
+});
